requirements.txt
opencv-python>=4.8.0
ultralytics>=8.0.0
face_recognition>=1.3.0
deep-sort-realtime>=1.3.2
streamlit>=1.30.0
python-dotenv>=1.0.0
twilio>=8.5.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
plotly>=5.18.0
